requirements.txt
MetaTrader5
pandas
numpy
scikit-learn
tensorflow
python-dotenv
PyYAML
fastapi
uvicorn[standard]
tqdm
lightgbm
xgboost
joblib
